do not crash if a name is unknown
this can be caused because of, e.g.:
extern PFN_vkCreateDevice vkCreateDevice;
declarations like this do not have a FunctionDecl associated with them,
so we don't know what they are.

also, there may be some empty names. i do not know what they are.
my guess is that they're lambdas in c++.